Move egress proxy token internals into service
Canonical egress proxy token API:

    manager, err := egressproxy.NewTokenManager(secret)
    token, err := manager.MintToken(egressproxy.TokenRequest{
        PluginName: "github",
        SessionID: "session-123",
        AllowedHosts: []string{"api.github.com"},
        DefaultAction: egress.PolicyDeny,
    })
    target, err := manager.ResolveToken(token)

Compatibility API remains available:

    manager, err := providerhost.NewEgressProxyTokenManager(secret)
    token, err := manager.MintToken(providerhost.EgressProxyTokenRequest{PluginName: "github"})
